requirements.txt
pandas
pymongo
Flask-SQLAlchemy
textblob
finnhub-python
psycopg2-binary